fix(kleos-cli): store importance as u8 0-10 to match server schema
CLI was sending importance as f32 0.0-1.0 but server expects i32 0-10,
causing HTTP 422 "invalid type: floating point, expected i32" on every
store call. Aligning to server schema.

+++ kleos-cli/src/main.rs
@@ -30,9 +30,9 @@ enum Commands {
         /// Category (task, discovery, decision, state, issue, general, reference)
         #[arg(short, long, default_value = "general")]
         category: String,
-        /// Importance score 0.0-1.0
+        /// Importance score 0-10 (integer)
         #[arg(short, long)]
-        importance: Option<f32>,
+        importance: Option<u8>,
         /// Comma-separated tags
         #[arg(short, long)]
         tags: Option<String>,
